ml-service/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.27.0
python-dotenv==1.0.1
scikit-learn==1.8.0
numpy>=1.26.4
pandas>=2.2.2
vaderSentiment==3.3.2
cachetools==5.3.3
pydantic==2.10.6